sync with prod

+++ server/player/tenhou/management/commands/recalculate_tenhou_accounts.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 from time import sleep
 
 from django.core.management.base import BaseCommand
@@ -20,9 +19,19 @@ def get_date_string():
 class Command(BaseCommand):
     def add_arguments(self, parser):
         parser.add_argument("--nickname", default=None, type=str)
+        parser.add_argument("--rebuild-from-zero", default=False, type=bool)
+        parser.add_argument("--with-curl", default=False, type=bool)
+
+    def is_need_update(self, now, tenhou_object) -> bool:
+        if tenhou_object.last_recalculated_date is not None:
+            return now.toordinal() - tenhou_object.last_recalculated_date.toordinal() > 0
+        else:
+            return True
 
     def handle(self, *args, **options):
         tenhou_nickname = options.get("nickname")
+        rebuild_from_zero = options.get("rebuild_from_zero")
+        with_pycurl = options.get("with_curl")
         print("{0}: Start".format(get_date_string()))
 
         if tenhou_nickname:
@@ -31,23 +40,30 @@ class Command(BaseCommand):
             tenhou_objects = TenhouNickname.active_objects.all()
         tenhou_players_count = tenhou_objects.count()
         current_player_index = 1
+        now = timezone.now()
         for tenhou_object in tenhou_objects:
             print(f"[{current_player_index}/{tenhou_players_count}] Processing {tenhou_object.tenhou_username}")
+            if rebuild_from_zero or (rebuild_from_zero is False and self.is_need_update(now, tenhou_object)):
+                TenhouGameLog.objects.filter(tenhou_object=tenhou_object).delete()
 
-            TenhouGameLog.objects.filter(tenhou_object=tenhou_object).delete()
+                player_games, account_start_date, four_players_rate, is_active_account = (
+                    download_all_games_from_nodochi(
+                        tenhou_object.tenhou_username, only_ranking_games=True, with_pycurl=with_pycurl
+                    )
+                )
 
-            player_games, account_start_date, four_players_rate = download_all_games_from_nodochi(
-                tenhou_object.tenhou_username, only_ranking_games=True
-            )
+                if is_active_account:
+                    tenhou_object.username_created_at = account_start_date
+                    save_played_games(tenhou_object, player_games)
+                    recalculate_tenhou_statistics_for_four_players(tenhou_object, player_games, four_players_rate, now)
+                else:
+                    tenhou_object.is_active = False
+                    tenhou_object.is_main = False
+                    tenhou_object.save()
 
-            tenhou_object.username_created_at = account_start_date
-
-            save_played_games(tenhou_object, player_games)
-
-            recalculate_tenhou_statistics_for_four_players(tenhou_object, player_games, four_players_rate)
+                # let's be gentle and don't ddos nodochi
+                sleep(10)
+            else:
+                print("already updated, skipped...")
             current_player_index = current_player_index + 1
-
-            # let's be gentle and don't ddos nodochi
-            sleep(10)
-
         print("{0}: End".format(get_date_string()))
